drop possibly redundant sqrt(K) factor from violation checking
When checking the strong (active?) and rest (inactive?) sets during
gradient descent, it compares norms to lambda * m[g] * sqrt(K[g]). For
the group lasso, by default m[g] (the group multiplier) is set to
sqrt(K[g]), so this has the effect of incorrectly checking the
violation.

This fix might break other penalties, but seems to improve the
behavior of the group lasso