support variable SVD truncation thresholds
Add svd.thresh argument to orthogonalize (and set default to 1e-10,
which is what it used to be hard-coded to). This allows the user to
specify how large singular values must be in on order to be preserved
in the truncated SVD used for group orthogonalization. Note: this
changes the default behavior a little bit. Currently, singular values
are just compared directly to 1e-10. However, because the columns of X
are scaled to have norm sqrt(n) (and not unit norm), as n grows, the
singular values will grow with sqrt(n). The new behavior is to compare
singular values to sqrt(n) * svd.thresh. Under the old behavior, if n
was sufficiently (extremely) large, it could cause some singular
values to be retained, even if they are effectively very small.

In order to expose this functionality in grpreg (and elsewhere), newXG
now passes `...` to orthogonalize, and grpreg passes `...` to `newXG`.

Also, added a note to the documentation for grpreg discussing the use
of `...` notation for this. Perhaps other user-facing functions should
have similar documentation changes, but I only use grpreg.

Finally, added a test to the standardization-ortho suite of tests
which includes a nearly rank-deficient design and tests if different
values of svd.thresh produce the expected behavior. We hard code the
seed here for reproducibility of the test (it's unlikely but possible
that an unusual draw could lead to inconsistent thresholding behavior).